+++ client/types/account/account-details.d.ts
@@ -6,7 +6,7 @@ export interface AccountDetailsData {
 	payout_status: {
 		text: string;
 		background_color: 'green' | 'yellow' | 'red';
-		icon: 'published' | 'caution' | 'error';
+		icon?: 'published' | 'caution' | 'error';
 		popover?: {
 			text: string;
 			cta_text: string;
@@ -18,7 +18,7 @@ export interface AccountDetailsData {
 		background_color: 'yellow' | 'red';
 		cta_text?: string;
 		cta_link?: string;
-		icon: 'caution';
+		icon?: 'caution';
 	} | null;
 }
